+++ lizard_progress/process_uploaded_file.py
@@ -13,61 +13,16 @@ from __future__ import division
 import logging
 import os
 import shutil
-import time
 import traceback
 
 from django.db import transaction
-from django.conf import settings
 
 from lizard_progress import models
 from lizard_progress import specifics
+from lizard_progress.util import directories
 from lizard_progress.tools import unique_filename
 
 logger = logging.getLogger(__name__)
-
-
-def document_root():
-    """Get the document root for uploaded files as an absolute path.
-    If LIZARD_PROGRESS_ROOT is given in settings, return that,
-    otherwise the directory var/lizard_progress/ under BUILDOUT_DIR.
-    """
-
-    root = getattr(settings, 'LIZARD_PROGRESS_ROOT', None)
-    if root is None:
-        root = os.path.join(settings.BUILDOUT_DIR,
-                            'var', 'lizard_progress')
-    return root
-
-
-def make_uploaded_file_path(root, project, contractor,
-                            measurement_type, filename):
-    """Gives the path to some uploaded file, which depends on the
-    project it is for, the contractor that uploaded it and the
-    measurement type that got its data from this file.
-
-    Project, contractor, measurement_type can each be either a
-    model instance of that type or a string containing the slug
-    of one.
-
-    Can be used both for absolute file paths (pass in document_root()
-    as root) or for URLs that will be passed to Nginx for X-Sendfile
-    (uses /protected/ as the root).
-
-    External URLs should use a reverse() call to the
-    lizard_progress_filedownload view instead of this function."""
-
-    if isinstance(project, models.Project):
-        project = project.slug
-    if isinstance(contractor, models.Contractor):
-        contractor = contractor.slug
-    if isinstance(measurement_type, models.MeasurementType):
-        measurement_type = measurement_type.slug
-
-    return os.path.join(root,
-                        project,
-                        contractor,
-                        measurement_type,
-                        os.path.basename(filename))
 
 
 def process_uploaded_file(uploaded_file_id):
@@ -232,14 +187,8 @@ def path_for_uploaded_file(uploaded_file, measurement_type):
     """Create dirname based on project etc. Guaranteed not to
     exist yet at the time of checking."""
 
-    dirname = os.path.dirname(make_uploaded_file_path(
-            document_root(),
-            uploaded_file.project, uploaded_file.contractor,
-            measurement_type, 'dummy'))
-
-    # Create directory if does not exist yet
-    if not os.path.exists(dirname):
-        os.makedirs(dirname)
+    dirname = directories.measurement_type_dir(
+        uploaded_file.project, uploaded_file.contractor, measurement_type)
 
     # Figure out a filename that doesn't exist yet
     orig_filename = os.path.basename(uploaded_file.filename)