+++ lizard_progress/util/directories.py
@@ -1,0 +1,148 @@
+# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
+# -*- coding: utf-8 -*-
+
+"""Helper functions to find out where a given file should be put /
+retrieved from."""
+
+# Python 3 is coming
+from __future__ import division
+
+import os
+
+from django.conf import settings
+
+
+BASE_DIR = getattr(
+    settings,
+    'LIZARD_PROGRESS_ROOT',
+    os.path.join(settings.BUILDOUT_DIR, 'var', 'lizard_progress'))
+
+
+def mk(directory):
+    """Create directory if it doesn't exist yet, then return it."""
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+    return directory
+
+
+def project_dir(project):
+    """Return base directory for this project."""
+    return mk(os.path.join(
+            BASE_DIR,
+            project.organization.name,
+            project.slug))
+
+
+def contractor_dir(project, contractor):
+    """Return base directory for this contractor in this project."""
+    return mk(os.path.join(
+            project_dir(project),
+            contractor.slug))
+
+
+def measurement_type_dir(project, contractor, measurement_type):
+    """Return base directory for this measurement type in this project
+    for this contractor."""
+    return mk(os.path.join(
+            contractor_dir(project, contractor),
+            measurement_type.slug))
+
+
+def make_uploaded_file_path(root, project, contractor,
+                            measurement_type, filename):
+    """Gives the path to some uploaded file, which depends on the
+    project it is for, the contractor that uploaded it and the
+    measurement type that got its data from this file.
+
+    Project, contractor, measurement_type can each be either a
+    model instance of that type or a string containing the slug
+    of one.
+
+    Can be used both for absolute file paths (pass in
+    directories.BASE_DIR as root) or for URLs that will be passed to
+    Nginx for X-Sendfile (uses /protected/ as the root).
+
+    External URLs should use a reverse() call to the
+    lizard_progress_filedownload view instead of this function."""
+
+    project = getattr(project, 'slug', project)
+    contractor = getattr(contractor, 'slug', contractor)
+    measurement_type = getattr(measurement_type, 'slug', measurement_type)
+
+    return os.path.join(root,
+                        project,
+                        contractor,
+                        measurement_type,
+                        os.path.basename(filename))
+
+
+def results_dir(project, contractor):
+    """Directory where scripts put result files for this contractor in
+    this project."""
+    return mk(os.path.join(
+            contractor_dir(project, contractor),
+            'final_results'))
+
+
+def exports_dir(project, contractor):
+    """Directory where scripts put result files for this contractor in
+    this project."""
+    return mk(os.path.join(
+            contractor_dir(project, contractor),
+            'export'))
+
+
+def reports_dir(project, contractor):
+    """Directory where uploads put reports from this contractor in
+    this project."""
+    return mk(os.path.join(
+            contractor_dir(project, contractor),
+            'reports'))
+
+
+def shapefile_dir(project, contractor):
+    """Directory where uploads put shapefiles from this contractor in
+    this project."""
+    return mk(os.path.join(
+            contractor_dir(project, contractor),
+            'shapefile'))
+
+
+def location_shapefile_dir(project, contractor):
+    """Directory where wizards put location shapefiles from this
+    contractor in this project."""
+    return mk(os.path.join(
+            contractor_dir(project, contractor),
+            'locations'))
+
+
+def hydrovakken_dir(project):
+    """Directory where wizards put hydrovakken shapefiles for this project."""
+    return mk(os.path.join(
+            project_dir(project),
+            'hydrovakken'))
+
+
+def organization_files_dir(organization):
+    return mk(os.path.join(
+            BASE_DIR,
+            organization.name,
+            'files'))
+
+
+def files_in(d):
+    for f in os.listdir(d):
+        if os.path.isfile(os.path.join(d, f)):
+            yield os.path.join(d, f)
+
+
+def human_size(path):
+    size = os.stat(path).st_size
+
+    if size < 1000:
+        return "{0} bytes".format(size)
+
+    if size < 1000000:
+        return "{0:.1f}KB".format(size / 1000)
+
+    return "{0:.1f}MB".format(size / 1000000)